pyproject.toml
[project]
name = "nebulabot"
version = "0.2.0"
description = "A modular AI chatbot framework inspired by AstrBot"
authors = [
    { name = "Antigravity", email = "assistant@example.com" }
]
dependencies = [
    "fastapi",
    "uvicorn",
    "openai",
    "anthropic",
    "google-generativeai",
    "python-dotenv",
    "pydantic",
    "httpx"
]
requires-python = ">=3.11"

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.ruff]
line-length = 88
target-version = "py311"
